+++ src/wai/json/object/property/_BoolProperty.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 from ...schema import BOOL_SCHEMA
-from .._typing import PropertyValueType, Absent
+from .._typing import PropertyValueType, Absent, OptionallyPresent
 from ._RawProperty import RawProperty
 
 
@@ -13,7 +13,7 @@ class BoolProperty(RawProperty):
                  name: Optional[str] = None,
                  *,
                  optional: bool = False,
-                 default: PropertyValueType = Absent):
+                 default: OptionallyPresent[PropertyValueType] = Absent):
         super().__init__(
             name,
             schema=BOOL_SCHEMA,